moved extraction to new folder "image processing"
